Move ReifyWindows etc to DirectRunner, don't subclass from GBK
